requirements.txt
pillow==10.3.0
pygame==2.5.2